requirements.txt
pytest
fastapi
python-dotenv
azure-storage-blob
aiohttp
pytest-asyncio
tox
